fix: match hosted cancellation signals
Amp-Thread-ID: https://ampcode.com/threads/T-019f8cdf-78cd-763a-84b1-feb48d60d8f4

Amp-Thread-ID: https://ampcode.com/threads/T-019f8cdf-78cd-763a-84b1-feb48d60d8f4

+++ internal/runtime/process_unix.go
@@ -13,13 +13,24 @@ func configureProcessGroup(command *exec.Cmd) {
 	command.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 }
 
-func terminateProcessGroup(ctx context.Context, pid int, grace time.Duration, finished <-chan struct{}) {
+func terminateProcessGroup(ctx context.Context, pid int, interruptGrace, terminateGrace time.Duration, finished <-chan struct{}) {
 	select {
 	case <-ctx.Done():
-		_ = syscall.Kill(-pid, syscall.SIGTERM)
+		_ = syscall.Kill(-pid, syscall.SIGINT)
 	case <-finished:
 		return
 	}
+	if waitForProcessGroupExit(pid, interruptGrace) {
+		return
+	}
+	_ = syscall.Kill(-pid, syscall.SIGTERM)
+	if waitForProcessGroupExit(pid, terminateGrace) {
+		return
+	}
+	_ = syscall.Kill(-pid, syscall.SIGKILL)
+}
+
+func waitForProcessGroupExit(pid int, grace time.Duration) bool {
 	timer := time.NewTimer(grace)
 	defer timer.Stop()
 	ticker := time.NewTicker(10 * time.Millisecond)
@@ -27,11 +38,10 @@ func terminateProcessGroup(ctx context.Context, pid int, grace time.Duration, fi
 	for {
 		select {
 		case <-timer.C:
-			_ = syscall.Kill(-pid, syscall.SIGKILL)
-			return
+			return false
 		case <-ticker.C:
 			if err := syscall.Kill(-pid, 0); err == syscall.ESRCH {
-				return
+				return true
 			}
 		}
 	}
